+++ internal/cli/atlas/customdbroles/custom_db_roles_test.go
@@ -20,6 +20,8 @@ import (
 	"testing"
 
 	"github.com/mongodb/mongocli/internal/test"
+	"github.com/stretchr/testify/assert"
+	"go.mongodb.org/atlas/mongodbatlas"
 )
 
 func TestBuilder(t *testing.T) {
@@ -29,4 +31,262 @@ func TestBuilder(t *testing.T) {
 		5,
 		[]string{},
 	)
+}
+
+func Test_appendActions(t *testing.T) {
+	type args struct {
+		existingActions []mongodbatlas.Action
+		newActions      []mongodbatlas.Action
+	}
+	tests := []struct {
+		name string
+		args args
+		want []mongodbatlas.Action
+	}{
+		{
+			name: "empty",
+			args: args{
+				existingActions: []mongodbatlas.Action{},
+				newActions:      []mongodbatlas.Action{},
+			},
+			want: []mongodbatlas.Action{},
+		},
+		{
+			name: "no new actions",
+			args: args{
+				existingActions: []mongodbatlas.Action{
+					{
+						Action: "TEST",
+						Resources: []mongodbatlas.Resource{
+							{
+								Collection: "test",
+								Db:         "test",
+							},
+						},
+					},
+				},
+				newActions: []mongodbatlas.Action{},
+			},
+			want: []mongodbatlas.Action{
+				{
+					Action: "TEST",
+					Resources: []mongodbatlas.Resource{
+						{
+							Collection: "test",
+							Db:         "test",
+						},
+					},
+				},
+			},
+		},
+		{
+			name: "different actions",
+			args: args{
+				existingActions: []mongodbatlas.Action{
+					{
+						Action: "TEST",
+						Resources: []mongodbatlas.Resource{
+							{
+								Collection: "test",
+								Db:         "test",
+							},
+						},
+					},
+				},
+				newActions: []mongodbatlas.Action{
+					{
+						Action: "NEW",
+						Resources: []mongodbatlas.Resource{
+							{
+								Collection: "test",
+								Db:         "test",
+							},
+						},
+					},
+				},
+			},
+			want: []mongodbatlas.Action{
+				{
+					Action: "TEST",
+					Resources: []mongodbatlas.Resource{
+						{
+							Collection: "test",
+							Db:         "test",
+						},
+					},
+				},
+				{
+					Action: "NEW",
+					Resources: []mongodbatlas.Resource{
+						{
+							Collection: "test",
+							Db:         "test",
+						},
+					},
+				},
+			},
+		},
+		{
+			name: "merge",
+			args: args{
+				existingActions: []mongodbatlas.Action{
+					{
+						Action: "TEST",
+						Resources: []mongodbatlas.Resource{
+							{
+								Collection: "test",
+								Db:         "test2",
+							},
+						},
+					},
+				},
+				newActions: []mongodbatlas.Action{
+					{
+						Action: "TEST",
+						Resources: []mongodbatlas.Resource{
+							{
+								Collection: "test",
+								Db:         "test",
+							},
+						},
+					},
+				},
+			},
+			want: []mongodbatlas.Action{
+				{
+					Action: "TEST",
+					Resources: []mongodbatlas.Resource{
+						{
+							Collection: "test",
+							Db:         "test",
+						},
+						{
+							Collection: "test",
+							Db:         "test2",
+						},
+					},
+				},
+			},
+		},
+	}
+	for _, tt := range tests {
+		args := tt.args
+		want := tt.want
+		t.Run(tt.name, func(t *testing.T) {
+			got := appendActions(args.existingActions, args.newActions)
+			assert.ElementsMatch(t, got, want)
+		})
+	}
+}
+
+func Test_joinActions(t *testing.T) {
+	type args struct {
+		newActions []mongodbatlas.Action
+	}
+	tests := []struct {
+		name string
+		args args
+		want []mongodbatlas.Action
+	}{
+		{
+			name: "empty",
+			args: args{
+				newActions: []mongodbatlas.Action{},
+			},
+			want: []mongodbatlas.Action{},
+		},
+		{
+			name: "no duplicate",
+			args: args{
+				newActions: []mongodbatlas.Action{
+					{
+						Action: "TEST",
+						Resources: []mongodbatlas.Resource{
+							{
+								Collection: "test",
+								Db:         "test",
+							},
+						},
+					},
+					{
+						Action: "TEST2",
+						Resources: []mongodbatlas.Resource{
+							{
+								Collection: "test",
+								Db:         "test",
+							},
+						},
+					},
+				},
+			},
+			want: []mongodbatlas.Action{
+				{
+					Action: "TEST",
+					Resources: []mongodbatlas.Resource{
+						{
+							Collection: "test",
+							Db:         "test",
+						},
+					},
+				},
+				{
+					Action: "TEST2",
+					Resources: []mongodbatlas.Resource{
+						{
+							Collection: "test",
+							Db:         "test",
+						},
+					},
+				},
+			},
+		},
+		{
+			name: "duplicates",
+			args: args{
+				newActions: []mongodbatlas.Action{
+					{
+						Action: "TEST",
+						Resources: []mongodbatlas.Resource{
+							{
+								Collection: "test",
+								Db:         "test",
+							},
+						},
+					},
+					{
+						Action: "TEST",
+						Resources: []mongodbatlas.Resource{
+							{
+								Collection: "test",
+								Db:         "test1",
+							},
+						},
+					},
+				},
+			},
+			want: []mongodbatlas.Action{
+				{
+					Action: "TEST",
+					Resources: []mongodbatlas.Resource{
+						{
+							Collection: "test",
+							Db:         "test1",
+						},
+						{
+							Collection: "test",
+							Db:         "test",
+						},
+					},
+				},
+			},
+		},
+	}
+	for _, tt := range tests {
+		args := tt.args
+		want := tt.want
+		t.Run(tt.name, func(t *testing.T) {
+			got := joinActions(args.newActions)
+			assert.ElementsMatch(t, got, want)
+		})
+	}
 }
