+++ internal/cli/atlas/customdbroles/update_test.go
@@ -23,7 +23,6 @@ import (
 	"github.com/mongodb/mongocli/internal/flag"
 	"github.com/mongodb/mongocli/internal/mocks"
 	"github.com/mongodb/mongocli/internal/test"
-	"github.com/stretchr/testify/assert"
 	"go.mongodb.org/atlas/mongodbatlas"
 )
 
@@ -72,150 +71,4 @@ func TestUpdateBuilder(t *testing.T) {
 		0,
 		[]string{flag.ProjectID, flag.Output, flag.Privilege, flag.InheritedRole, flag.Append},
 	)
-}
-
-func Test_appendActions(t *testing.T) {
-	type args struct {
-		existingActions []mongodbatlas.Action
-		newActions      []mongodbatlas.Action
-	}
-	tests := []struct {
-		name string
-		args args
-		want []mongodbatlas.Action
-	}{
-		{
-			name: "empty",
-			args: args{
-				existingActions: []mongodbatlas.Action{},
-				newActions:      []mongodbatlas.Action{},
-			},
-			want: []mongodbatlas.Action{},
-		},
-		{
-			name: "no new actions",
-			args: args{
-				existingActions: []mongodbatlas.Action{
-					{
-						Action: "TEST",
-						Resources: []mongodbatlas.Resource{
-							{
-								Collection: "test",
-								Db:         "test",
-							},
-						},
-					},
-				},
-				newActions: []mongodbatlas.Action{},
-			},
-			want: []mongodbatlas.Action{
-				{
-					Action: "TEST",
-					Resources: []mongodbatlas.Resource{
-						{
-							Collection: "test",
-							Db:         "test",
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "different actions",
-			args: args{
-				existingActions: []mongodbatlas.Action{
-					{
-						Action: "TEST",
-						Resources: []mongodbatlas.Resource{
-							{
-								Collection: "test",
-								Db:         "test",
-							},
-						},
-					},
-				},
-				newActions: []mongodbatlas.Action{
-					{
-						Action: "NEW",
-						Resources: []mongodbatlas.Resource{
-							{
-								Collection: "test",
-								Db:         "test",
-							},
-						},
-					},
-				},
-			},
-			want: []mongodbatlas.Action{
-				{
-					Action: "TEST",
-					Resources: []mongodbatlas.Resource{
-						{
-							Collection: "test",
-							Db:         "test",
-						},
-					},
-				},
-				{
-					Action: "NEW",
-					Resources: []mongodbatlas.Resource{
-						{
-							Collection: "test",
-							Db:         "test",
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "merge",
-			args: args{
-				existingActions: []mongodbatlas.Action{
-					{
-						Action: "TEST",
-						Resources: []mongodbatlas.Resource{
-							{
-								Collection: "test",
-								Db:         "test2",
-							},
-						},
-					},
-				},
-				newActions: []mongodbatlas.Action{
-					{
-						Action: "TEST",
-						Resources: []mongodbatlas.Resource{
-							{
-								Collection: "test",
-								Db:         "test",
-							},
-						},
-					},
-				},
-			},
-			want: []mongodbatlas.Action{
-				{
-					Action: "TEST",
-					Resources: []mongodbatlas.Resource{
-						{
-							Collection: "test",
-							Db:         "test",
-						},
-						{
-							Collection: "test",
-							Db:         "test2",
-						},
-					},
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		args := tt.args
-		want := tt.want
-		t.Run(tt.name, func(t *testing.T) {
-			got := appendActions(args.existingActions, args.newActions)
-			assert.ElementsMatch(t, got, want)
-		})
-	}
 }
